feat: manage initiative (#11)
* feat: add settings for the encounter tracker

* feat: add initiative dialog

* chore: change initiative to double

* feat: add autoroll initiative behavior on encounter start

* feat: change combatants list to be reorderable

* style: format

+++ lib/features/combatant/models/combatant.dart
@@ -12,7 +12,7 @@ class Combatant extends Equatable {
   final String name;
   final int currentHp;
   final int maxHp;
-  final int initiative;
+  final double initiative;
   final int armorClass;
   final int initiativeModifier;
   final int? level;
@@ -47,7 +47,7 @@ class Combatant extends Equatable {
     String? name,
     int? currentHp,
     int? maxHp,
-    int? initiative,
+    double? initiative,
     int? armorClass,
     int? initiativeModifier,
     int? level,
